looped pali word ingestion mostly done

+++ db/migrate/20230731174054_create_looped_pali_words.rb
@@ -1,7 +1,7 @@
 class CreateLoopedPaliWords < ActiveRecord::Migration[7.0]
   def change
     create_table :looped_pali_words do |t|
-      t.integer :index
+      t.bigint :index
       t.string :original_pali
       t.string :original_url
       t.string :pali
